+++ storage-resize-images/__tests__/util.test.ts
@@ -1,4 +1,5 @@
-import { extractFileNameWithoutExtension, startsWithArray } from "../functions/src/util";
+import { extractFileNameWithoutExtension } from "../functions/src/util";
+
 describe("extractFileNameWithoutExtension", () => {
   const filePathWithExtension = "/ref/to/my/image.png";
   const filePathWithoutExtension = "/ref/to/my/image";
@@ -6,11 +7,11 @@ describe("extractFileNameWithoutExtension", () => {
 
   describe("when file path includes file extension", () => {
     it("extracts the basename without extension", () => {
-      const fileNameWithoutExtension = extractFileNameWithoutExtension(
+      const fileNameWithExtension = extractFileNameWithoutExtension(
         filePathWithExtension,
         ext
       );
-      expect(fileNameWithoutExtension).toBe("image");
+      expect(fileNameWithExtension).toBe("image");
     });
   });
 
@@ -21,38 +22,6 @@ describe("extractFileNameWithoutExtension", () => {
         ext
       );
       expect(fileNameWithoutExtension).toBe("image");
-    });
-  });
-});
-
-describe("startsWithArray", () => {
-  const absolutePathList = [
-    "/users",
-    "/posts/pictures",
-    "/design/resources/images",
-  ];
-  const filePathDirnamePost = "/posts/pictures";
-  const filePathDirnameUser = "/users/avatars";
-  const filePathDirnameIcon = "/icons";
-
-  describe("when file path dirname is included in the absolute paths at same level", () => {
-    it("check the file path dirname agains absolute path list", () => {
-      const result = startsWithArray(absolutePathList, filePathDirnamePost);
-      expect(result).toBeTruthy();
-    });
-  });
-
-  describe("when file path dirname is included in the absolute paths not at same level", () => {
-    it("check the file path dirname agains absolute path list", () => {
-      const result = startsWithArray(absolutePathList, filePathDirnameUser);
-      expect(result).toBeTruthy();
-    });
-  });
-
-  describe("when file path dirname is not included in the absolute paths", () => {
-    it("check the file path dirname agains absolute path list", () => {
-      const result = startsWithArray(absolutePathList, filePathDirnameIcon);
-      expect(result).toBeFalsy();
     });
   });
 });