+++ storage-resize-images/functions/src/config.ts
@@ -22,4 +22,5 @@ export default {
   includePathList: process.env.INCLUDE_PATH_LIST,
   excludePathList: process.env.EXCLUDE_PATH_LIST,
   deleteOriginalFile: process.env.DELETE_ORIGINAL_FILE === "true",
+  imageType: process.env.IMAGE_TYPE,
 };